Move wrapper test to file with other tests
